% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcfpca.R
\name{gc_fpca}
\alias{gc_fpca}
\title{Generalized Conditional Functional Principal Components Analysis (GC-FPCA)}
\usage{
gc_fpca(
  formula,
  data,
  binwidth = 10,
  family = "gaussian",
  pve = NULL,
  npc = NULL,
  periodicity = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula specifying the GLMM model.}

\item{data}{A data frame with columns: \code{id} (subject), \code{index} (time or index), and \code{Y} (outcome variable).}

\item{binwidth}{Numeric; controls the width of the bins for local fitting. Defaults to 10.}

\item{family}{A family object for specifying the distribution and link function for the GLMM. Defaults to \code{binomial}.}

\item{pve}{Numeric; proportion of variance explained, used to choose the number of principal components unless \code{npc} is specified.}

\item{npc}{Numeric; number of smooth PCs to extract. If \code{NULL}, \code{npc} is chosen based on \code{pve}.}

\item{...}{Additional arguments passed to the \code{mgcv::bam} function.}
}
\value{
An object of class \code{gc_fpca} containing:
\item{betaHat}{Estimated fixed effects coefficient functions}
\item{betaHat.sd}{Estimated fixed effects coefficient standard deviation along the functional domain}
\item{efunctions}{\eqn{K \times npc} matrix of estimated FPC basis functions.}
\item{scores}{\eqn{I \times npc} matrix of estimated FPC scores.}
\item{mu}{Estimated population-level mean function.}
\item{Yhat}{FPC approximation of subject-specific means.}
\item{family}{The family used for GLMM.}
}
\description{
This is the main function for the GCFPCA package.
}
\examples{
# Simulate data and fit GC-FPCA
df_list <- gen_data(N = 100, J = 50, run_num = 1)
gcfpca_mod <- gc_fpca(df_list, bin_width = 10, family = binomial())
}
