% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcfpca_simu.R
\name{gcfpca_simu}
\alias{gcfpca_simu}
\title{Simulate longitudinal functional response Y (matrix) and a scalar predictor X (vector)}
\usage{
gcfpca_simu(
  family = "gaussian",
  I = 100,
  K = 100,
  beta0_true,
  beta1_true,
  fe_case = 1,
  re_case = 1
)
}
\arguments{
\item{family}{distribution of longitudinal functional data, including "gaussian", "binomial", "poisson".}

\item{I}{number of subjects.}

\item{K}{number of grid points on the functional domain.}

\item{beta0_true}{true intercept function.}

\item{beta1_true}{true fixed effect function.}

\item{fe_case}{Takes on values of 1 or 2. If case = 1, true fixed effect
coefficient is generated from a Bernoulli(0.5) distribution .
If case = 2 true fixed effect, coefficient is from a N(0,1) distribution.}

\item{re_case}{Takes on values of 1 or 2. If case = 1 true eigenfunctions
are based on alternating sine and cosines. If case = 2 true eigenfunctions are
based on sqrt functions.}
}
\value{
a data frame containing generated predictors and longitudinal functional outcomes.
}
\description{
Simulate longitudinal functional response Y (matrix) and a scalar predictor X (vector)
}
